fix: поправил замечания, но не меняя саму длинную функцию
